free-chatgpt-accounts/requirements.txt
pandas
tabulate
OpenAIAuth
retry
PyGithub
schedule